LPS-96442 Inline JSPackageAdapter.
